refactor: split files under 1000 lines (Phase 5)
cli_calc_functions.c (1311 -> 849):
  - Extracted binary operator evaluations to cli_calc_binops.c

CLIcore_help.c (1466 -> 754):
  - Extracted help_command and command_info_search to CLIcore_help_command.c

ZernikePolyn.c (1017 -> 705):
  - Extracted get_zer, get_zerns, fit_zer, etc. to ZernikePolyn_fit.c
